+++ somni-template/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 - Specify template style options in `---`-fenced frontmatter
 - `#replace "SOMETHING" with expression` blocks
+- `#include "path"` blocks
 
 # [0.3.2] - 2026-07-15